+++ flake.nix
@@ -9,7 +9,7 @@
 	outputs = inputs@{ flake-parts, ... }:
 		flake-parts.lib.mkFlake { inherit inputs; } {
 			systems = [ "x86_64-linux" "aarch64-linux" "aarch64-darwin" "x86_64-darwin" ];
-			perSystem = { self', pkgs, system, ... }: {
+			perSystem = { self', pkgs, lib, system, ... }: {
 				packages = let pkgName = "fish-subAbbr"; in {
 					default = self'.packages.${pkgName};
 					${pkgName} = builtins.derivation {
@@ -21,6 +21,19 @@
 						HOME = "/tmp/"; # Temporary home directory to write history to
 						PATH = pkgs.coreutils+"/bin";
 					};
+				};
+				devShells = {
+					default = self'.devShells.pkg;
+					pkg = pkgs.mkShellNoCC {
+						shellHook = ''exec ${lib.meta.getExe pkgs.fish} --init-command='
+							functions --erase -- sub-abbr sub-abbrs (functions --all | string match --entire --regex -- ^_sub-abbr_) (functions --all | string match --entire --regex -- ^__sub_2D_abbr__expand_)
+							for script in ${self'.packages.default}/share/fish/*/*.fish
+								source {$script}
+							end
+							sub-abbr identity erase (sub-abbr identity list)'
+						'';
+					};
+					# Also a shell that directly sources the scripts in the current directory
 				};
 			};
 		};